Initial commit
Auto heal sort of works.  If you are busy, such as in combat, it does not interrupt current action.  I still need to figure that out.
Looting spells you don't know mostly works, but it will only pick up first spell you don't know from a corpse.
My first version of auto loot sort of worked but the timing was all wrong.  The version committed has the scaffolding for a finite state machine.  I plan to convert the code currently in timer.cs to the fsm class.  I think that is going to work better.